Moved ArrayList containing all Moves into Move class for SRP
